Add files via upload
This is the clean version of the Skeleton program, you can duplicate this branch to create specific answers to individual problems.

The master branch will include a copy of ALL changes that work, make sure to test your changes on a new branch before committing them to the master copy. All commits to the master copy should be commented to explain the code and what it does.  You can add your name in the comments so everyone know who did the hard work.